#628 added and styled header and banner
